requirements.txt
discord
sqlalchemy
aiosqlite
asyncpg
requests
python-dotenv